requirements.txt
Django==1.11.10
djangorestframework==3.7.7
olefile==0.45.1
pytz==2017.3
whitenoise==3.2.2
